+++ src/pages/DecorateBouquetPage/components/index.ts
@@ -1,3 +1,4 @@
 export { default as Ribbons } from './Ribbons'
 export { default as FlowerFrame } from './FlowerFrame'
 export { default as TabContents } from './TabContents'
+export { default as BouquetFrame } from './BouquetFrame'
